+++ gpu/src/wgpu_impl/draw_color_triangles_pass.rs
@@ -103,6 +103,7 @@ impl DrawColorTrianglesPass {
               },
             ],
           }],
+          compilation_options: Default::default(),
         },
         fragment: Some(wgpu::FragmentState {
           module: &self.shader,
@@ -112,6 +113,7 @@ impl DrawColorTrianglesPass {
             blend: Some(wgpu::BlendState::ALPHA_BLENDING),
             write_mask: wgpu::ColorWrites::all(),
           })],
+          compilation_options: Default::default(),
         }),
         primitive: wgpu::PrimitiveState {
           topology: wgpu::PrimitiveTopology::TriangleList,