+++ CHANGELOG.md
@@ -36,6 +36,8 @@ Please only add new entries below the [Unreleased](#unreleased---releasedate) he
 - **macros**: polish the compile error message of invalid filed in `@$var {}` (#556 @M-Adoo)
 - **gpu**: Removed dependency on the texture array feature of wgpu. (#562, @M-Adoo)
 - **algo**: removed `Resource` and rename `ShareResource` to `Resource`. (#564, @M-Adoo)
+- **dev-helper**: Support specific the comparison of image tests. (#573 @M-Adoo)
+- **dev-helper**: If test images differ, both actual and difference images are saved with the expected image. (#573 @M-Adoo)
 
 ### Documented